+++ .gitignore
@@ -193,7 +193,7 @@
 /test-fdset
 /test-fileio
 /test-fstab-util
-/test-fw-util
+/test-firewall-util
 /test-hashmap
 /test-hostname
 /test-icmp6-rs
