+++ src/test/test-firewall-util.c
@@ -20,7 +20,7 @@
 ***/
 
 #include "log.h"
-#include "fw-util.h"
+#include "firewall-util.h"
 
 #define MAKE_IN_ADDR_UNION(a,b,c,d) (union in_addr_union) { .in.s_addr = htobe32((uint32_t) (a) << 24 | (uint32_t) (b) << 16 | (uint32_t) (c) << 8 | (uint32_t) (d))}
